Formatting: Don't convert smilies in ignored tags that have attributes.
Smilies in ignored tags are not supposed to be converted to emoji, but this can malfunction if the tag has attributes. For example, the Preformatted block with add a `class` to the `<pre>` tag.

Props pento, jikamens.
Merges [45569] to the 5.2 branch.
Fixes #47489.
Built from https://develop.svn.wordpress.org/branches/5.2@45837