Accessibility: Improve and modernize user interface controls: Revert the new links focus style.
Thew new dotted outline for the links focus style introduced in [46241] doesn't appear to be ready to guarantee a good indication of focus.
It was agreed to restore the previous links focus style and postpone exploration for a new style to the next release cycle.
Partially reverts [46241] and [46293].

See #34904, #47153.

Built from https://develop.svn.wordpress.org/trunk@46425